+++ app/ui/settings.py
@@ -33,8 +33,8 @@ def render_settings() -> GenerateSettings:
             "gpt-3.5-turbo-0301",
             "gpt-4",
             "gpt-4-0314",
-            "gpt-4-32k",
-            "gpt-4-32k-0314",
+            # "gpt-4-32k",
+            # "gpt-4-32k-0314",
         ]
         model_ids_sorted = sorted(filtered_list)
         default_model_index = model_ids_sorted.index(config["DEFAULT_GPT_MODEL"])
