+++ pyproject.toml
@@ -14,13 +14,13 @@ tiktoken = "^0.2.0"
 streamlit = "^1.18.1"
 pyrate-limiter = "^2.10.0"
 praw = "^7.7.0"
+colorlog = "^6.7.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
 mypy = "^0.991"
 types-requests = "^2.28.11.8"
 black = "^23.1.0"
-colorlog = "^6.7.0"
 debugpy = "^1.6.6"
 pylint = "^2.16.2"
